+++ Trains3D/Engine/Geometry/EGFigure.h
@@ -12,57 +12,57 @@
 @interface EGLine : NSObject
 + (id)line;
 - (id)init;
-+ (EGLine*)newWithSlope:(float)slope point:(EGPoint)point;
++ (EGLine*)newWithSlope:(CGFloat)slope point:(EGPoint)point;
 + (EGLine*)newWithP1:(EGPoint)p1 p2:(EGPoint)p2;
-+ (float)calculateSlopeWithP1:(EGPoint)p1 p2:(EGPoint)p2;
-+ (float)calculateConstantWithSlope:(float)slope point:(EGPoint)point;
++ (CGFloat)calculateSlopeWithP1:(EGPoint)p1 p2:(EGPoint)p2;
++ (CGFloat)calculateConstantWithSlope:(CGFloat)slope point:(EGPoint)point;
 - (BOOL)containsPoint:(EGPoint)point;
 - (BOOL)isVertical;
 - (BOOL)isHorizontal;
 - (id)intersectionWithLine:(EGLine*)line;
-- (float)xIntersectionWithLine:(EGLine*)line;
+- (CGFloat)xIntersectionWithLine:(EGLine*)line;
 - (BOOL)isRightPoint:(EGPoint)point;
-- (float)slope;
-- (id)moveWithDistance:(float)distance;
-- (float)angle;
-- (float)degreeAngle;
+- (CGFloat)slope;
+- (id)moveWithDistance:(CGFloat)distance;
+- (CGFloat)angle;
+- (CGFloat)degreeAngle;
 - (EGLine*)perpendicularWithPoint:(EGPoint)point;
 @end
 
 
 @interface EGSlopeLine : EGLine
-@property (nonatomic, readonly) float slope;
-@property (nonatomic, readonly) float constant;
+@property (nonatomic, readonly) CGFloat slope;
+@property (nonatomic, readonly) CGFloat constant;
 
-+ (id)slopeLineWithSlope:(float)slope constant:(float)constant;
-- (id)initWithSlope:(float)slope constant:(float)constant;
++ (id)slopeLineWithSlope:(CGFloat)slope constant:(CGFloat)constant;
+- (id)initWithSlope:(CGFloat)slope constant:(CGFloat)constant;
 - (BOOL)containsPoint:(EGPoint)point;
 - (BOOL)isVertical;
 - (BOOL)isHorizontal;
-- (float)xIntersectionWithLine:(EGLine*)line;
-- (float)yForX:(float)x;
+- (CGFloat)xIntersectionWithLine:(EGLine*)line;
+- (CGFloat)yForX:(CGFloat)x;
 - (id)intersectionWithLine:(EGLine*)line;
 - (BOOL)isRightPoint:(EGPoint)point;
-- (id)moveWithDistance:(float)distance;
-- (float)angle;
+- (id)moveWithDistance:(CGFloat)distance;
+- (CGFloat)angle;
 - (EGLine*)perpendicularWithPoint:(EGPoint)point;
 @end
 
 
 @interface EGVerticalLine : EGLine
-@property (nonatomic, readonly) float x;
+@property (nonatomic, readonly) CGFloat x;
 
-+ (id)verticalLineWithX:(float)x;
-- (id)initWithX:(float)x;
++ (id)verticalLineWithX:(CGFloat)x;
+- (id)initWithX:(CGFloat)x;
 - (BOOL)containsPoint:(EGPoint)point;
 - (BOOL)isVertical;
 - (BOOL)isHorizontal;
-- (float)xIntersectionWithLine:(EGLine*)line;
+- (CGFloat)xIntersectionWithLine:(EGLine*)line;
 - (id)intersectionWithLine:(EGLine*)line;
 - (BOOL)isRightPoint:(EGPoint)point;
-- (float)slope;
-- (id)moveWithDistance:(float)distance;
-- (float)angle;
+- (CGFloat)slope;
+- (id)moveWithDistance:(CGFloat)distance;
+- (CGFloat)angle;
 - (EGLine*)perpendicularWithPoint:(EGPoint)point;
 @end
 
@@ -81,7 +81,7 @@
 + (id)lineSegmentWithP1:(EGPoint)p1 p2:(EGPoint)p2;
 - (id)initWithP1:(EGPoint)p1 p2:(EGPoint)p2;
 + (EGLineSegment*)newWithP1:(EGPoint)p1 p2:(EGPoint)p2;
-+ (EGLineSegment*)newWithX1:(float)x1 y1:(float)y1 x2:(float)x2 y2:(float)y2;
++ (EGLineSegment*)newWithX1:(CGFloat)x1 y1:(CGFloat)y1 x2:(CGFloat)x2 y2:(CGFloat)y2;
 - (BOOL)isVertical;
 - (BOOL)isHorizontal;
 - (EGLine*)line;
@@ -91,7 +91,7 @@
 - (BOOL)endingsContainPoint:(EGPoint)point;
 - (id<CNList>)segments;
 - (EGLineSegment*)moveWithPoint:(EGPoint)point;
-- (EGLineSegment*)moveWithX:(float)x y:(float)y;
+- (EGLineSegment*)moveWithX:(CGFloat)x y:(CGFloat)y;
 @end
 
 
@@ -107,11 +107,11 @@
 
 @interface EGThickLineSegment : NSObject<EGFigure>
 @property (nonatomic, readonly) EGLineSegment* segment;
-@property (nonatomic, readonly) float thickness;
-@property (nonatomic, readonly) float thickness_2;
+@property (nonatomic, readonly) CGFloat thickness;
+@property (nonatomic, readonly) CGFloat thickness_2;
 
-+ (id)thickLineSegmentWithSegment:(EGLineSegment*)segment thickness:(float)thickness;
-- (id)initWithSegment:(EGLineSegment*)segment thickness:(float)thickness;
++ (id)thickLineSegmentWithSegment:(EGLineSegment*)segment thickness:(CGFloat)thickness;
+- (id)initWithSegment:(EGLineSegment*)segment thickness:(CGFloat)thickness;
 - (EGRect)boundingRect;
 - (id<CNList>)segments;
 @end
